Add files via upload
Make improvements to all files. Code optimization, breaking down codes into sections for better understanding and readability. Added additional error handling and improved the UI. I added a function to allow the end user the ability to direct the Excel and Word files to be saved to their desired locations.